Factor out the function and a struct to crew_cabin.rs
